+++ CHANGELOG.md
@@ -51,6 +51,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `valueFactory`, `build`, `builder`, `standard` in `org.partiql.lang.CompilerPipeline`
 - **Breaking**: Removes deprecated `org.partiql.lang.eval.visitors.GroupKeyReferencesVisitorTransform`
 
+- **Breaking**: Removes `org.partiql.lang.mappers.StaticTypeMapper`
+- **Breaking**: Removes `org.partiql.lang.mappers.IonSchemaMapper`
+- **Breaking**: Removes `org.partiql.lang.mappers.TypeNotFoundException`
+- **Breaking**: Removes `org.partiql.lang.mappers.getBaseTypeName()`
 - **Breaking**: Removes unused/deprecated enums `KEYWORD`, `TOKEN_TYPE`, `EXPECTED_TOKEN_TYPE`, `EXPECTED_TOKEN_TYPE_1_OF_2`,
   `EXPECTED_TOKEN_TYPE_2_OF_2`, `TIMESTAMP_STRING`, `NARY_OP` from `org.partiql.lang.errors.Property`
 - **Breaking**: Removes unused `tokenTypeValue()` from `org.partiql.lang.errors.PropertyValue`